+++ ChangeLog.md
@@ -8,6 +8,7 @@
   - Peeper
   - Diving Kit
 - Configurable random contract upon typing ``contract`` rather than interacting with the terminal application associated with contracts.
+- Implemented "Sleight Of Hand" tier upgrade which increases the reload speed of the shotgun item.
 
 </details>